requirements.txt
gradio==5.9.1
lmdb==1.3.0
numpy==1.24.4
onnx==1.13.0
onnxmltools==1.11.1
onnxruntime_gpu==1.13.1
Pillow==11.0.0
six==1.17.0
tensorrt==10.7.0
tensorrt_cu12==10.7.0
timm==1.0.12
torch==1.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.67.1
